refactor: Rename `lib/appUids.js`
