+++ lib/app-uids.test.js
@@ -5,7 +5,7 @@ const proxyquire = require('proxyquire');
 const sinon = require('sinon');
 
 const getAppStub = sinon.stub().resolves({ appUid: 'AUID', tenantUid: 'OUID' });
-const appUids = proxyquire('./appUids', {
+const appUids = proxyquire('./app-uids', {
   './clientUtils': {
     getPlatformClientWithAccessKey: () => ({
       apps: {
