+++ lib/plugin.js
@@ -11,7 +11,7 @@ const wrap = require('./wrap');
 const injectLogsIamRole = require('./injectLogsIamRole');
 const injectOutputOutputs = require('./injectOutputOutputs');
 const wrapClean = require('./wrapClean');
-const getAppUids = require('./appUids');
+const getAppUids = require('./app-uids');
 const removeDestination = require('./removeDestination');
 const { saveDeployment, createAndSetDeploymentUid } = require('./deployment');
 const variables = require('./variables');
