+++ lib/plugin.test.js
@@ -62,7 +62,7 @@ describe('plugin', () => {
         }),
       },
       '@serverless/platform-client': platformClientStub,
-      './appUids': () => ({ appUid: '000000000000000000', orgUid: '000000000000000000' }),
+      './app-uids': () => ({ appUid: '000000000000000000', orgUid: '000000000000000000' }),
     });
 
     // REMOVING GETPROVIDREMOCK() AND LOGMOCK() AND USING THESLS INSTANCE BELOW
